Disable unstable sort.
After upgrading to rust 1.78, tests for the unstable sorting algorithm
are failing.  I think I found at least one actual bug in the
implementation (not sure why it only got triggered now), but there are
more, since tests are still crashing or returning wrong results.

I moved the unstable sort algorithm to a separate module and disabled it
for now.  In fact I didn't realize it was used in the first place.  It
was not well tested and was meant to be disabled, but apparently I did
not disable it everywhere.
