+++ crates/dbsp/src/utils/mod.rs
@@ -8,6 +8,8 @@ mod tuple;
 #[cfg(test)]
 mod vec_ext;
 
+use std::cmp::Ordering;
+use std::fmt::Debug;
 use std::{fmt::Display, hint::unreachable_unchecked};
 
 pub use advance_retreat::{
@@ -22,12 +24,27 @@ pub use consolidation::{
 pub use consolidation::consolidate_pairs;
 
 pub use sample::sample_slice;
-pub use sort::{stable_sort, stable_sort_by, unstable_sort, unstable_sort_by};
+pub use sort::{stable_sort, stable_sort_by};
 pub use tuple::{
     ArchivedTup0, ArchivedTup1, ArchivedTup10, ArchivedTup2, ArchivedTup3, ArchivedTup4,
     ArchivedTup5, ArchivedTup6, ArchivedTup7, ArchivedTup8, ArchivedTup9, Tup0, Tup1, Tup10, Tup2,
     Tup3, Tup4, Tup5, Tup6, Tup7, Tup8, Tup9,
 };
+
+// mod unstable_sort;
+
+// FIXME: unstable sort implementation is currently broken; use stable sorting instead.
+pub fn unstable_sort<T: Ord + Debug>(slice: &mut [T]) {
+    stable_sort(slice)
+}
+
+// FIXME: unstable sort implementation is currently broken; use stable sorting instead.
+pub fn unstable_sort_by<T, F>(slice: &mut [T], cmp: F)
+where
+    F: Fn(&T, &T) -> Ordering,
+{
+    stable_sort_by(slice, cmp)
+}
 
 #[cfg(test)]
 pub(crate) use vec_ext::VecExt;
